requirements.txt
mexc-api
python-dotenv
telethon
python-binance
kucoin-futures-python
kucoin-python
ccxt
bitmart-python-sdk-api